fake_news_detection/requirements-ml.txt
# Core ML / DL
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
xgboost>=2.0.0
numpy>=1.26.0
pandas>=2.1.0

# Text preprocessing
nltk>=3.8.0
sastrawi>=1.0.1            # Indonesian stemmer
Unidecode>=1.3.8

# OCR for image uploads
easyocr>=1.7.1
Pillow>=10.0.0

# Cross-reference scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Visualisation & notebook
matplotlib>=3.8.0
seaborn>=0.13.0
jupyterlab>=4.1.0
ipywidgets>=8.1.0

# User-facing app
streamlit>=1.33.0

# Reporting
reportlab>=4.1.0
